requirements.txt
Django==4.0.4
django-tinymce==3.4.0
Pillow==9.1.1

